# perf: using generics instead of reflect for better performance