--- a/pkg/planner/core/issuetest/BUILD.bazel
+++ b/pkg/planner/core/issuetest/BUILD.bazel
@@ -9,7 +9,6 @@ go_test(
     ],
     data = glob(["testdata/**"]),
     flaky = True,
-    race = "on",
     shard_count = 12,
     deps = [
         "//pkg/parser",