--- a/pkg/util/slice/slice_test.go
+++ b/pkg/util/slice/slice_test.go
@@ -23,22 +23,18 @@ import (
 
 func TestSlice(t *testing.T) {
 	tests := []struct {
-		a      []int
-		anyOf  bool
-		noneOf bool
-		allOf  bool
+		a     []int
+		allOf bool
 	}{
-		{[]int{}, false, true, true},
-		{[]int{1, 2, 3}, true, false, false},
-		{[]int{1, 3}, false, true, false},
-		{[]int{2, 2, 4}, true, false, true},
+		{[]int{}, true},
+		{[]int{1, 2, 3}, false},
+		{[]int{1, 3}, false},
+		{[]int{2, 2, 4}, true},
 	}
 
 	for _, test := range tests {
 		t.Run(fmt.Sprint(test.a), func(t *testing.T) {
-			even := func(i int) bool { return test.a[i]%2 == 0 }
-			require.Equal(t, test.anyOf, AnyOf(test.a, even))
-			require.Equal(t, test.noneOf, NoneOf(test.a, even))
+			even := func(val int) bool { return val%2 == 0 }
 			require.Equal(t, test.allOf, AllOf(test.a, even))
 		})
 	}