<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

as diff I run some bench mark local seems it can have a better performance, 
and I am still learning it, not sure it is 100% right


benchmark:

```cli
BenchmarkAnyOf_New_Small-10             144057651                8.305 ns/op           0 B/op          0 allocs/op
BenchmarkAnyOf_Old_Small-10             121768545                9.815 ns/op           0 B/op          0 allocs/op
BenchmarkAnyOf_New_Medium-10            30066207                39.44 ns/op            0 B/op          0 allocs/op
BenchmarkAnyOf_Old_Medium-10            29319988                40.62 ns/op            0 B/op          0 allocs/op
BenchmarkAnyOf_New_Large-10             30374875                39.85 ns/op            0 B/op          0 allocs/op
BenchmarkAnyOf_Old_Large-10             29228607                40.93 ns/op            0 B/op          0 allocs/op
BenchmarkAnyOf_New_XLarge-10            29980574                39.67 ns/op            0 B/op          0 allocs/op
BenchmarkAnyOf_Old_XLarge-10            29309665                40.75 ns/op            0 B/op          0 allocs/op
BenchmarkAnyOf_New_WorstCase-10          1713400               701.5 ns/op             0 B/op          0 allocs/op
BenchmarkAnyOf_Old_WorstCase-10          1734120               690.2 ns/op             0 B/op          0 allocs/op
BenchmarkAnyOf_New_BestCase-10          1000000000               1.168 ns/op           0 B/op          0 allocs/op
BenchmarkAnyOf_Old_BestCase-10          554618565                2.169 ns/op           0 B/op          0 allocs/op
BenchmarkAnyOf_StringSlice_New-10        9882049               121.1 ns/op             0 B/op          0 allocs/op
BenchmarkAnyOf_StringSlice_Old-10        3419708               350.7 ns/op             0 B/op          0 allocs/op
BenchmarkAnyOf_Float64Slice_New-10       9883820               121.4 ns/op             0 B/op          0 allocs/op
BenchmarkAnyOf_Float64Slice_Old-10       3424080               352.5 ns/op             0 B/op          0 allocs/op

BenchmarkNoneOf_New_Medium-10           30326326                39.30 ns/op            0 B/op          0 allocs/op
BenchmarkNoneOf_Old_Medium-10           29606510                40.30 ns/op            0 B/op          0 allocs/op
BenchmarkNoneOf_New_Large-10            30235838                39.68 ns/op            0 B/op          0 allocs/op
BenchmarkNoneOf_Old_Large-10            29523350                40.33 ns/op            0 B/op          0 allocs/op


BenchmarkAllOf_New_Medium-10            593807331                1.847 ns/op           0 B/op          0 allocs/op
BenchmarkAllOf_Old_Medium-10            277587498                4.333 ns/op           0 B/op          0 allocs/op
BenchmarkAllOf_New_Large-10             651542288                1.846 ns/op           0 B/op          0 allocs/op
BenchmarkAllOf_Old_Large-10             275393266                4.340 ns/op           0 B/op          0 allocs/op
BenchmarkAllOf_New_WorstCase-10         1000000000               1.148 ns/op           0 B/op          0 allocs/op
BenchmarkAllOf_Old_WorstCase-10         344983638                3.495 ns/op           0 B/op          0 allocs/op
BenchmarkAllOf_New_BestCase-10           1715023               710.6 ns/op             0 B/op          0 allocs/op
BenchmarkAllOf_Old_BestCase-10           1000000              1186 ns/op      
```

code
```go
// Copyright 2021 PingCAP, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

package slice

import (
	"reflect"
	"testing"
)

// oldAnyOf is the old reflection-based implementation for comparison
func oldAnyOf(s any, p func(int) bool) bool {
	l := reflect.ValueOf(s).Len()
	for i := range l {
		if p(i) {
			return true
		}
	}
	return false
}

// oldNoneOf is the old reflection-based implementation for comparison
func oldNoneOf(s any, p func(int) bool) bool {
	return !oldAnyOf(s, p)
}

// oldAllOf is the old reflection-based implementation for comparison
func oldAllOf(s any, p func(int) bool) bool {
	np := func(i int) bool {
		return !p(i)
	}
	return oldNoneOf(s, np)
}

// Benchmark data
var (
	intSlice10    = []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
	intSlice100   = make([]int, 100)
	intSlice1000  = make([]int, 1000)
	intSlice10000 = make([]int, 10000)
)

func init() {
	for i := range intSlice100 {
		intSlice100[i] = i
	}
	for i := range intSlice1000 {
		intSlice1000[i] = i
	}
	for i := range intSlice10000 {
		intSlice10000[i] = i
	}
}

// Predicate functions for benchmarking
var (
	// Always false predicate - worst case for AnyOf, best case for AllOf
	alwaysFalse = func(int) bool { return false }
	// Always true predicate - best case for AnyOf, worst case for AllOf
	alwaysTrue = func(int) bool { return true }
	// Find middle element - average case
	findMiddle = func(i int) bool { return i == 50 }
	// Even numbers predicate
	isEven = func(i int) bool { return i%2 == 0 }
)

// AnyOf benchmarks
func BenchmarkAnyOf_New_Small(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AnyOf(intSlice10, findMiddle)
	}
}

func BenchmarkAnyOf_Old_Small(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAnyOf(intSlice10, findMiddle)
	}
}

func BenchmarkAnyOf_New_Medium(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AnyOf(intSlice100, findMiddle)
	}
}

func BenchmarkAnyOf_Old_Medium(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAnyOf(intSlice100, findMiddle)
	}
}

func BenchmarkAnyOf_New_Large(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AnyOf(intSlice1000, findMiddle)
	}
}

func BenchmarkAnyOf_Old_Large(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAnyOf(intSlice1000, findMiddle)
	}
}

func BenchmarkAnyOf_New_XLarge(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AnyOf(intSlice10000, findMiddle)
	}
}

func BenchmarkAnyOf_Old_XLarge(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAnyOf(intSlice10000, findMiddle)
	}
}

// AnyOf worst case (always false)
func BenchmarkAnyOf_New_WorstCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AnyOf(intSlice1000, alwaysFalse)
	}
}

func BenchmarkAnyOf_Old_WorstCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAnyOf(intSlice1000, alwaysFalse)
	}
}

// AnyOf best case (always true)
func BenchmarkAnyOf_New_BestCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AnyOf(intSlice1000, alwaysTrue)
	}
}

func BenchmarkAnyOf_Old_BestCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAnyOf(intSlice1000, alwaysTrue)
	}
}

// NoneOf benchmarks
func BenchmarkNoneOf_New_Medium(b *testing.B) {
	for i := 0; i < b.N; i++ {
		NoneOf(intSlice100, findMiddle)
	}
}

func BenchmarkNoneOf_Old_Medium(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldNoneOf(intSlice100, findMiddle)
	}
}

func BenchmarkNoneOf_New_Large(b *testing.B) {
	for i := 0; i < b.N; i++ {
		NoneOf(intSlice1000, findMiddle)
	}
}

func BenchmarkNoneOf_Old_Large(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldNoneOf(intSlice1000, findMiddle)
	}
}

// AllOf benchmarks
func BenchmarkAllOf_New_Medium(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AllOf(intSlice100, isEven)
	}
}

func BenchmarkAllOf_Old_Medium(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAllOf(intSlice100, isEven)
	}
}

func BenchmarkAllOf_New_Large(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AllOf(intSlice1000, isEven)
	}
}

func BenchmarkAllOf_Old_Large(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAllOf(intSlice1000, isEven)
	}
}

// AllOf worst case (always false predicate)
func BenchmarkAllOf_New_WorstCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AllOf(intSlice1000, alwaysFalse)
	}
}

func BenchmarkAllOf_Old_WorstCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAllOf(intSlice1000, alwaysFalse)
	}
}

// AllOf best case (always true predicate)
func BenchmarkAllOf_New_BestCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		AllOf(intSlice1000, alwaysTrue)
	}
}

func BenchmarkAllOf_Old_BestCase(b *testing.B) {
	for i := 0; i < b.N; i++ {
		oldAllOf(intSlice1000, alwaysTrue)
	}
}

// Different slice types to test generics vs reflection
func BenchmarkAnyOf_StringSlice_New(b *testing.B) {
	stringSlice := make([]string, 1000)
	for i := range stringSlice {
		stringSlice[i] = "test"
	}
	
	for i := 0; i < b.N; i++ {
		AnyOf(stringSlice, func(i int) bool { return i == 500 })
	}
}

func BenchmarkAnyOf_StringSlice_Old(b *testing.B) {
	stringSlice := make([]string, 1000)
	for i := range stringSlice {
		stringSlice[i] = "test"
	}
	
	for i := 0; i < b.N; i++ {
		oldAnyOf(stringSlice, func(i int) bool { return i == 500 })
	}
}

func BenchmarkAnyOf_Float64Slice_New(b *testing.B) {
	floatSlice := make([]float64, 1000)
	for i := range floatSlice {
		floatSlice[i] = float64(i)
	}
	
	for i := 0; i < b.N; i++ {
		AnyOf(floatSlice, func(i int) bool { return i == 500 })
	}
}

func BenchmarkAnyOf_Float64Slice_Old(b *testing.B) {
	floatSlice := make([]float64, 1000)
	for i := range floatSlice {
		floatSlice[i] = float64(i)
	}
	
	for i := 0; i < b.N; i++ {
		oldAnyOf(floatSlice, func(i int) bool { return i == 500 })
	}
}

```

<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #62972

Problem Summary:

### What changed and how does it work?

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```